chore: bump version from 0.3.2 to 0.3.3 refactor: create benchmark subpackage to encapsulate different benchmarking types (currently only pgvector) refactor: move
