+++ chem_mrl/benchmark/__init__.py
@@ -1,0 +1,5 @@
+from __future__ import annotations
+
+from .DatabaseBenchmark import PgVectorBenchmark
+
+__all__ = ["PgVectorBenchmark"]
