+++ pyproject.toml
@@ -1,6 +1,6 @@
 [project]
 name = "chem-mrl"
-version = "0.3.2"
+version = "0.3.3"
 description = "SMILES-based Matryoshka Representation Learning Embeddings"
 license = { text = "Apache 2.0" }
 readme = "README.md"
@@ -10,20 +10,20 @@ authors = [
 maintainers = [
     { name = "Emmanuel Cortes", email = "ecortes@berkeley.com" },
 ]
-requires-python = ">=3.11"
+requires-python = ">=3.10"
 
 dependencies = [
     # ML dependencies
     "sentence-transformers==3.4.1",
-    "transformers==4.48.1", 
+    "transformers>=4.34.0", 
     "optuna==4.2.0",
     "wandb==0.19.4",
     "torch>=2.0.0",
     "ninja==1.11.1.3",
     "scikit-learn==1.6.1",
-    "pandas==2.2.2",
+    "pandas>=2.0.0",
     "pandarallel==1.6.5",
-    "numpy==2.2.2",
+    "numpy>=1.26.4",
 ]
 
 [project.urls]
